Add second implementation of existing karatsuba function.
This one uses the product-of-differences form of the karatsuba
equations. The previous one treated the oversized inner multiply
as a nested karatsuba multiplication and then optimized it
into requiring no more than the single multiply that was already
required for the inner multiply (because a single bit can multiply
numbers using logical operations, and does not need a full
multiply instruction).